UIとビジネスロジックを分離する
